Update flake to support nix 2.8
nix 2.7 renamed defaultApp and defaultPackage. Both the
old and new names are supported in 2.7, but 2.8 has removed
support for the old names, breaking the nix run invocation.

Old names are kept in this PR to keep compatibility with nix 2.6,
but could be removed if support of this version is not needed anymore.