allow non-flake builds with flake-enabled nix
Adds a --file option, which forces the non-flake build logic to always
run, regardless of whether or not the current nix is flake-enabled.

This allows using deploy-rs with repositories which do not use flakes
at all but use an external source-managing tool such as niv or npins
(the previous behaviour could only really support flake-based
repositories with a flake-compat fallback defined in default.nix).